cleanup: some google readability clang-tidy warnings - mostly braces

+++ src/intraNonBonded/intraNonBondedMap.cpp
@@ -126,7 +126,7 @@ std::pair<double, double> IntraNonBondedMap::calculateSingleInteraction(
     auto coulombEnergy    = 0.0;
     auto nonCoulombEnergy = 0.0;
 
-    const auto atomIdx2 = size_t(::abs(atomIndex2AsInt));
+    const auto atomIdx2 = static_cast<size_t>(::abs(atomIndex2AsInt));
     const bool scale    = atomIndex2AsInt < 0;
 
     const auto &pos1 = _molecule->getAtomPosition(atomIdx1);
@@ -141,7 +141,7 @@ std::pair<double, double> IntraNonBondedMap::calculateSingleInteraction(
 
     if (distance < CoulombPotential::getCoulombRadiusCutOff())
     {
-        const auto charge1 = _molecule->getPartialCharge(size_t(atomIdx1));
+        const auto charge1 = _molecule->getPartialCharge(atomIdx1);
         const auto charge2 = _molecule->getPartialCharge(atomIdx2);
 
         const auto chargeProduct = charge1 * charge2;
